refactor: Improve ui experience.
